Implement full CRUD for cats and update cat model
Removed the previous cat model implementation because I did not need to
specify an initialize method in it.

Rolled back the previous DB migration because I forgot to write out the
name and age columns for the Cat DB table. They have now been properly
migrated.

Lastly, I wrote the full CRUD index, show, create, update, and destroy
actions for the Cats controller and enabled the routes for all actions.