Initializes owners controller and model
Generates a controller for Owners to which the cats will, in the future,
belong to.

Creates the Owner model and updates the migration file to store a name
and age for an owner instance.

Opens routes for all Owner actions despite them not yet being
implemented. In hindsight this was a bad idea as it makes sense to only
open the routes once the CRUD features have been created and tested.